Added two settings for Poggi's RM problem

+++ src/flow/initial_conditions/InitialConditions.cpp
@@ -832,7 +832,7 @@ InitialConditions::initializeDataOnPatch(
                                 double S = 0.0;
                                 for (int m = 40; m <= 66; m++)
                                 {
-                                    S += A*(cos(2*M_PI*m/0.05*x[1] + tan(m))*cos(2*M_PI*m/0.05*x_2 + tan(m)));
+                                    S += A*(cos(2*M_PI*m/0.05*x[1] + tan((double) m))*cos(2*M_PI*m/0.05*x_2 + tan((double) m)));
                                 }
                                 
                                 double phi = computeRandomModeLocationSettings1(x[1], x_2);
@@ -957,7 +957,7 @@ InitialConditions::initializeDataOnPatch(
                                 double S = 0.0;
                                 for (int m = 20; m <= 66; m++)
                                 {
-                                    S += A*(cos(2*M_PI*m/0.05*x[1] + tan(m))*cos(2*M_PI*m/0.05*x_2 + tan(m)));
+                                    S += A*(cos(2*M_PI*m/0.05*x[1] + tan((double) m))*cos(2*M_PI*m/0.05*x_2 + tan((double) m)));
                                 }
                                 
                                 double phi = computeRandomModeLocationSettings1(x[1], x_2);
@@ -1082,7 +1082,7 @@ InitialConditions::initializeDataOnPatch(
                                 double S = 0.0;
                                 for (int m = 40; m <= 66; m++)
                                 {
-                                    S += A*(cos(2*M_PI*m/0.05*x[1] + tan(3*m))*cos(2*M_PI*m/0.05*x_2 + tan(3*m)));
+                                    S += A*(cos(2*M_PI*m/0.05*x[1] + tan(3.0*m))*cos(2*M_PI*m/0.05*x_2 + tan(3.0*m)));
                                 }
                                 
                                 double phi = computeRandomModeLocationSettings1(x[1], x_2);
@@ -1207,7 +1207,7 @@ InitialConditions::initializeDataOnPatch(
                                 double S = 0.0;
                                 for (int m = 20; m <= 66; m++)
                                 {
-                                    S += A*(cos(2*M_PI*m/0.05*x[1] + tan(3*m))*cos(2*M_PI*m/0.05*x_2 + tan(3*m)));
+                                    S += A*(cos(2*M_PI*m/0.05*x[1] + tan(3.0*m))*cos(2*M_PI*m/0.05*x_2 + tan(3.0*m)));
                                 }
                                 
                                 double phi = computeRandomModeLocationSettings1(x[1], x_2);
